edura_core/requirements.txt
fastapi
uvicorn
python-dotenv
requests
aiohttp
